+++ js/constant.js
@@ -19,6 +19,7 @@ constant.ringAdjustAngleFactor = 3.6;
 constant.ringAdjustSizeFactor = 0.9;
 constant.ringMinRadiusSize = 108;
 
+constant.initView = -1;
 constant.radialView = 0;
 constant.listView = 1;
 constant.journalView = 2;
@@ -39,6 +40,7 @@ constant.listStepCount = 10;
 constant.listScrollLimit = 200;
 
 constant.timerUpdateNetwork = 1000;
+constant.timerBeforeTutorial = 500;
 constant.webAppType = 0; // thinClientType
 constant.appType = 1;  // clientType